refactor: promote compute_job to module
